+++ src/parser/monga_parser.c
@@ -144,7 +144,7 @@ union YYSTYPE
 {
 #line 17 "src/parser/monga.y" /* yacc.c:355  */
 
-  double              float_val;
+  float               float_val;
   int                 int_val;
   char               *string_val;
   char                char_val;
